Sneak / Store
Sneak now displays the first string of content, but the font is not cool
like it was, and its not based on time.  This still needs to be fixed.
Store has been updated so it now has if statements which govern which
level needs to come next.
Intro still goes straight to sneak for debugging purposes